requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pymysql
alembic
cryptography
python-multipart
passlib
python-dotenv
aiomysql
aiohttp
motor